SavedStateHandle in ViewState, updated Hilt
