+++ buildSrc/src/main/kotlin/Versions.kt
@@ -24,7 +24,7 @@ object Versions {
     const val constraintLayout = "2.0.3"
     const val lifeCycle = "2.2.0"
     const val dagger = "2.32"
-    const val hilt = "2.30.1-alpha"
+    const val hilt = "2.31.1-alpha"
     const val hiltJetpack = "1.0.0-alpha03"
     const val androidXEspresso = "3.3.0"
     const val navigation = "2.3.3"
