+++ example-hilt/src/main/java/app/futured/arkitekt/examplehilt/tools/SavedStateHandleExtension.kt
@@ -1,0 +1,6 @@
+package app.futured.arkitekt.examplehilt.tools
+
+import androidx.lifecycle.SavedStateHandle
+
+fun <T> SavedStateHandle.getOrThrow(key: String): T =
+    get<T>(key) ?: throw IllegalArgumentException("SavedStateHandle missing argument '$key'")
