--- example-hilt/src/main/java/app/futured/arkitekt/examplehilt/ui/MainActivityModule.kt
@@ -1,9 +1,0 @@
-package app.futured.arkitekt.examplehilt.ui
-
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityComponent
-
-@Module
-@InstallIn(ActivityComponent::class)
-class MainActivityModule
