+++ example-hilt/src/main/java/app/futured/arkitekt/examplehilt/ui/first/FirstEvent.kt
@@ -1,0 +1,6 @@
+package app.futured.arkitekt.examplehilt.ui.first
+
+import app.futured.arkitekt.core.event.Event
+
+sealed class FirstEvent : Event<FirstViewState>()
+data class NavigateToSecondFragmentEvent(val number: Int) : FirstEvent()
