+++ example-hilt/src/main/java/app/futured/arkitekt/examplehilt/ui/first/FirstView.kt
@@ -1,0 +1,5 @@
+package app.futured.arkitekt.examplehilt.ui.first
+
+import app.futured.arkitekt.core.BaseView
+
+interface FirstView : BaseView
