+++ example-hilt/src/main/java/app/futured/arkitekt/examplehilt/ui/second/SecondEvent.kt
@@ -1,0 +1,5 @@
+package app.futured.arkitekt.examplehilt.ui.second
+
+import app.futured.arkitekt.core.event.Event
+
+sealed class SecondEvent : Event<SecondViewState>()
