+++ example-hilt/src/main/java/app/futured/arkitekt/examplehilt/ui/second/SecondView.kt
@@ -1,0 +1,5 @@
+package app.futured.arkitekt.examplehilt.ui.second
+
+import app.futured.arkitekt.core.BaseView
+
+interface SecondView : BaseView
