+++ example-hilt/src/main/java/app/futured/arkitekt/examplehilt/ui/second/SecondViewModel.kt
@@ -1,0 +1,12 @@
+package app.futured.arkitekt.examplehilt.ui.second
+
+import app.futured.arkitekt.crusecases.BaseCrViewModel
+import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
+
+@HiltViewModel
+class SecondViewModel @Inject constructor(
+    override val viewState: SecondViewState
+) : BaseCrViewModel<SecondViewState>()
+
+
